% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emi.R
\name{cleanGXP}
\alias{cleanGXP}
\title{Clean GXP data}
\usage{
cleanGXP(df)
}
\arguments{
\item{df}{the data frame}
}
\description{
\code{cleanGXP} is called by getGxpData. Takes a wide form dataframe and returns
a cleaned data.table
}
\seealso{
Other emi: \code{\link{getGxpData}},
  \code{\link{loadGXPData}}, \code{\link{refreshGXPData}}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{emi}
