% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{codePeakPeriod}
\alias{codePeakPeriod}
\title{Add peak period to a data.table}
\usage{
codePeakPeriod(dt)
}
\arguments{
\item{dt}{the data table}
}
\description{
\code{codePeakPeriod} returns a dt with peakPeriod factor added. Assumes column
hms exists and is in hms form
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
