% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getFileList}
\alias{getFileList}
\title{Gets a list of files by path and pattern}
\usage{
getFileList(dPath, pattern)
}
\arguments{
\item{dPath}{the file path}

\item{pattern}{the pattern to match}
}
\description{
\code{getFileList} returns the list of files in a given path which match
a given pattern. Returns a data.table with 3 columns - the file name, it's
full path and it's file size in Mb
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
