% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emi.R
\name{getGxpData}
\alias{getGxpData}
\title{Download GXP data}
\usage{
getGxpData(f)
}
\arguments{
\item{years}{the years to get}
}
\description{
\code{getData} is called by refreshGXPData to get a single monthly NZ Electricity Authority Grid Exit Point data
from https://www.emi.ea.govt.nz/Wholesale/Datasets/Metered_data/Grid_export and
pre-processes them into monthly long form files with friendlier column names.
Function saves a long form data.table.
}
\seealso{
Other emi: 
\code{\link{cleanGXP}()},
\code{\link{loadGXPData}()},
\code{\link{refreshGXPData}()}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{emi}
