% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeca.R
\name{labelEECACircuits}
\alias{labelEECACircuits}
\title{Code the circuits}
\usage{
labelEECACircuits(dt, loadVar = "imputedTotalDemand_circuitsToSum_v1.1")
}
\arguments{
\item{dt}{the data table}

\item{loadVar}{the label for the total load 'circuit'. Default = \code{imputedTotalDemand_circuitsToSum_v1.1}}
}
\description{
\code{labelEECACircuits} expects a data.table with a \code{circuit} column and returns a data.table
with the circuit string split into circuitLabel (for easy string matching) and circuitID.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
