% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emi.R
\name{loadGXPData}
\alias{loadGXPData}
\title{Load gxp data}
\usage{
loadGXPData(files)
}
\arguments{
\item{files}{the gxp files}
}
\description{
\code{loadGXPData} loads monthly NZ Electricity Authority Grid Exit Point data
which has already been downloaded from https://www.emi.ea.govt.nz/Wholesale/Datasets/Metered_data/Grid_export
and and pre-processed into long form. Also requires a geo-coded look-up table
which is in the repo /data/input folder. Function returns a long form data.table.
}
\seealso{
Other emi: \code{\link{cleanGXP}},
  \code{\link{getGxpData}}, \code{\link{refreshGXPData}}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{emi}
