% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emi.R
\name{refreshGXPData}
\alias{refreshGXPData}
\title{Download , process and save monthly GXP data}
\usage{
refreshGXPData(years)
}
\arguments{
\item{years}{the years to get}
}
\description{
\code{refreshGXPData} gets monthly NZ Electricity Authority Grid Exit Point data
from https://www.emi.ea.govt.nz/Wholesale/Datasets/Metered_data/Grid_export and
pre-processes them into monthly long form files with friendlier column names.
Function saves a long form data.table.
}
\seealso{
Other emi: \code{\link{cleanGXP}},
  \code{\link{getGxpData}}, \code{\link{loadGXPData}}
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
\concept{emi}
